adapted HabitController to JWT

+++ spring/src/main/java/io/habitcare/web/controller/HabitController.java
@@ -1,12 +1,15 @@
 package io.habitcare.web.controller;
 
 import io.habitcare.web.dto.HabitDto;
+import io.habitcare.web.dto.CheckDto;
 import io.habitcare.web.model.Habit;
 import io.habitcare.web.service.habit.HabitService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import io.habitcare.web.service.jwt.JwtService;
+import io.habitcare.web.service.user.UserService;
 
 import java.util.List;
 
@@ -17,17 +20,27 @@ import static io.habitcare.web.mapper.HabitMapper.mapToHabitDto;
 @RequestMapping(path = "/api/habits")
 public class HabitController {
     private final HabitService habitService;
+    private final UserService userService;
+    private final JwtService jwtService;
 
     @Autowired
-    public HabitController(HabitService habitService) {
+    public HabitController(HabitService habitService, UserService userService, JwtService jwtService) {
         this.habitService = habitService;
+        this.userService = userService;
+        this.jwtService = jwtService;
     }
 
     @PostMapping(path = "/add")
-    public ResponseEntity<HabitDto> addHabit(@RequestBody HabitDto habitDto) {
+    public ResponseEntity<HabitDto> addHabit(@RequestBody HabitDto habitDto, @RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+
         Habit habit = mapFromHabitDto(habitDto);
         Habit savedHabit = habitService.createHabit(habit);
+        Long habitId = savedHabit.getId();
+        habitService.assignHabitToUser(userId,habitId);
         return new ResponseEntity<>(mapToHabitDto(savedHabit), HttpStatus.CREATED);
+
     }
 
     @GetMapping(path = "/get")
@@ -56,7 +69,56 @@ public class HabitController {
 
     @DeleteMapping(path = "/delete/{habitId}")
     public ResponseEntity<HabitDto> deleteHabit(@PathVariable Long habitId) {
+        if (!habitService.exists(habitId)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         habitService.deleteHabit(habitId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    @PostMapping(path = "/add-check/{habitId}")
+    public ResponseEntity<CheckDto> AddCheckHabit(@PathVariable Long habitId,@RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+        if(habitService.isTodayChecked(habitId, userId)) {
+            return new ResponseEntity<>(HttpStatus.LOCKED);
+        }
+        CheckDto checkDto = habitService.AddCheckHabit(habitId, userId);
+        return new ResponseEntity<>(checkDto, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping(path = "/remove-check/{habitId}")
+    public ResponseEntity<Void> removeCheckHabit(@PathVariable Long habitId, @RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+        if(!habitService.isTodayChecked(habitId, userId)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        habitService.removeCheckHabit(habitId, userId);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    @GetMapping(path = "/is-checked/{habitId}")
+    public ResponseEntity<Boolean> isTodayChecked(@PathVariable Long habitId, @RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+        boolean isChecked = habitService.isTodayChecked(habitId, userId);
+        return new ResponseEntity<>(isChecked, HttpStatus.OK);
+    }
+
+    @GetMapping(path = "/count-checks/{habitId}")
+    public ResponseEntity<Long> countChecks(@PathVariable Long habitId, @RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+        Long count = habitService.countChecks(habitId, userId);
+        return new ResponseEntity<>(count, HttpStatus.OK);
+    }
+
+    @GetMapping(path = "/streak/{habitId}")
+    public ResponseEntity<Long> getStreak(@PathVariable Long habitId, @RequestHeader("Authorization") String token) {
+        String email = jwtService.getEmailFromToken(token);
+        Long userId = userService.getUserIdByEmail(email);
+        Long streak = habitService.getStreak(habitId, userId);
+        return new ResponseEntity<>(streak, HttpStatus.OK);
     }
 }
